requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==42.0.1
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jsonfield==3.1.0
mysqlclient==2.1.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.10.0
urllib3==1.26.16
filetype==1.2.0
six==1.16.0
boto3==1.27.0
django-storages==1.13.2
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.1.13
openai==1.14.2
langchain-community==0.0.29
langchain-openai==0.1.0
tiktoken==0.6.0
faiss_cpu==1.8.0
django_advanced_password_validation==1.1.1
django-filter==23.2
pypdf==4.0.0
drf_yasg==1.21.7
celery==5.3.6
redis==5.0.2
selenium==4.18.1
webdriver-manager==4.0.1
pynput==1.7.6
pydparser==1.0.4
PyVirtualDisplay==3.0
pillow==10.2.0